fix: type casts and bytemuck derive for multi_brain module
- Cast vocab_size to usize in sample_with_temperature call
- Manual Pod/Zeroable impl for MessageGlyph ([f32; 60] too large for derive)
- Fix get_bind_group_layout call with explicit index 0
- Cast id params to usize for brains/regions vec lookups

+++ systems/infinite_map_rs/src/pixel_brain/infer.rs
@@ -1091,7 +1091,7 @@ impl PixelBrainInferencer {
             let next_token = if temperature > 0.0 {
                 // Sample with temperature
                 let logits = self.get_logits(current_token);
-                Self::sample_with_temperature(&logits, temperature, self.config.vocab_size)
+                Self::sample_with_temperature(&logits, temperature, self.config.vocab_size as usize)
             } else {
                 // Greedy (temperature = 0)
                 self.infer_token(current_token)
